Builds - Runtime issue where RX dataSize struct field is not updated - also keeping state 1 for ever

+++ GPD/plugin/rtos-support/soc/rtos_gpd.h
@@ -41,7 +41,7 @@ typedef enum gpd_ll_event_type {
 
 typedef struct gpd_ll_event_data
 {
-    uint32_t dataSize;
+    uint16_t dataSize;
     uint8_t *data;
 } gpd_ll_event_data_t;
